requirements.txt
discord.py~=2.4.0
typing~=4.12.2
openai~=0.28.0
aiohttp~=3.10.5
Flask~=3.0.3
responses
dotenv